backend/utils/access_control.py
from flask import request, jsonify, g
from functools import wraps
import pymysql
from models.user import User
from models.mentor_assignment import MentorAssignment
import jwt
import os
from dotenv import load_dotenv
import traceback

load_dotenv()
JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY', 'fallback_secret_key_for_development')

def get_token_from_header():
    """Extract token from Authorization header"""
    auth_header = request.headers.get('Authorization')
    if not auth_header:
        return None
    
    # Accept token with or without Bearer prefix
    if auth_header.startswith('Bearer '):
        return auth_header.split(' ')[1]
    return auth_header

def parse_token(token):
    """
    Simplified token parser that accepts any format
    and extracts user_id and role from it.
    """
    if not token:
        return None
    
    try:
        # First try to decode as standard JWT
        try:
            # Try decoding with the configured JWT secret key
            payload = jwt.decode(token, JWT_SECRET_KEY, algorithms=["HS256"])
            return {
                'user_id': payload.get('user_id', 1),
                'role': payload.get('role', 'admin'),
                'session_id': payload.get('session_id', 'default')
            }
        except jwt.InvalidTokenError:
            # JWT decoding failed, try custom formats
            pass
            
        # For token:user_id:role format
        if ':' in token:
            parts = token.split(':')
            
            # Try to extract user_id and role
            if len(parts) >= 3 and parts[0] == 'token':
                # Standard format: token:user_id:role[:session_id]
                return {
                    'user_id': int(parts[1]),
                    'role': parts[2],
                    'session_id': parts[3] if len(parts) > 3 else 'default'
                }
            elif len(parts) >= 2:
                # Simple format: user_id:role
                return {
                    'user_id': int(parts[0]),
                    'role': parts[1],
                    'session_id': 'default'
                }
        
        # Fallback: assume token is user_id and default to admin role
        return {
            'user_id': int(token) if token.isdigit() else 1,
            'role': 'admin',  # Default to admin for maximum access
            'session_id': 'default'
        }
    except Exception as e:
        print(f"Token parse error (using default): {str(e)}")
        traceback.print_exc()
        # If all parsing fails, return a default admin token for development
        return {
            'user_id': 1,
            'role': 'admin',
            'session_id': 'default'
        }

def verify_token(token=None):
    """Simplified token verification that skips database check"""
    if token is None:
        token = get_token_from_header()
    
    return parse_token(token)

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        if 'Authorization' in request.headers:
            auth_header = request.headers['Authorization']
            try:
                token = auth_header.split(" ")[1]
            except IndexError:
                # For development, use a default token
                print("WARNING: No token provided, using default admin token for development")
                token = "dev_token"
        
        if not token:
            return jsonify({'error': 'Authentication token is missing'}), 401
            
        try:
            # Try to decode as JWT first, then fallback to custom format
            data = parse_token(token)
            if not data:
                raise Exception("Invalid token format")
            
            # Set token data on request for convenience
            request.token_data = data
            
            # Set user on Flask g object for compatibility with @*_required decorators
            g.user = type('User', (), {'id': data['user_id'], 'role': data['role']})
            
            # Check if user exists in database
            user = User.query.get(data['user_id'])
            if not user:
                # As a fallback during development, allow token even if user doesn't exist
                print(f"WARNING: User with ID {data['user_id']} not found in database")
            
            return f(*args, **kwargs)
        except jwt.ExpiredSignatureError:
            return jsonify({'error': 'Token has expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'error': 'Invalid token'}), 401
        except Exception as e:
            traceback.print_exc()
            return jsonify({'error': f'Authentication error: {str(e)}'}), 401
            
    return decorated

def role_required(role):
    """Decorator factory for role checking"""
    def decorator(f):
        @wraps(f)
        @token_required
        def decorated_function(*args, **kwargs):
            # Check if the user has the required role
            if request.token_data['role'] != role:
                return jsonify({'error': f'{role.capitalize()} access required. Your role: {request.token_data["role"]}'}), 403
            return f(*args, **kwargs)
        return decorated_function
    return decorator

# Specific role decorators
admin_required = role_required('admin')
mentor_required = role_required('mentor')
student_required = role_required('student')

# Always allow access for development purposes
def check_student_access(student_id):
    """Always allow access in development mode"""
    return True 