backend/requirements.txt
# Flask and Extensions
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Login==0.5.0

# Database
PyMySQL==1.1.0
SQLAlchemy==1.4.23

# Security
Werkzeug==2.0.1
PyJWT==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0
email-validator==2.1.0.post1
pymysql==1.1.0

# Development Tools
pytest==6.2.5
black==21.9b0
flake8==3.9.2 